+++ spec/lib/capistrano/scm/git_spec.rb
@@ -27,6 +27,17 @@ module Capistrano
       Capistrano::Configuration.reset!
     end
 
+    describe "#set_defaults" do
+      it "sets ::git_wrapper_path in a shell-safe way" do
+        env.set(:tmp_dir, "/tmp")
+        env.set(:application, "my_app")
+        env.set(:stage, "staging")
+        env.set(:local_user, "(Git Web User) via ShipIt")
+        subject.set_defaults
+        expect(env.fetch(:git_wrapper_path)).to eq('/tmp/git-ssh-my_app-staging-\(Git\ Web\ User\)\ via\ ShipIt.sh')
+      end
+    end
+
     describe "#git" do
       it "should call execute git in the context, with arguments" do
         backend.expects(:execute).with(:git, :init)